my_robot: urdf: xacro: Add optical camera link
For a RGB-D camera in URDF file, we need to add an extra link and an
extra joint to the camera link in order to align the camera image in
Gazebo properly with the robot. The parent link of camera_optical_joint
should be properly configured to the original camera link.