my_robot: urdf: gazebo: Replace camera with kinnect RGBD one
Replace the normal camera with a Kinnect RGBD one and configure it in
this change.

The shared object file: libgazebo_ros_camera.so is replaced with that
of the Kinect shared object file: libgazebo_ros_openni_kinect.so.
Also, the <frameName> is updated to be the camera_link_optical link
created in the previous commit.

On top of this, additional parameters need to be set for the RGB-D
camera as well as matching the topics published by the drivers of its
real world counterpart.